+++ pebble2/src/main/main.h
@@ -1,0 +1,2 @@
+#pragma once
+#define REDRAW_INTERVAL_MS 1000
