+++ pebble2/wscript
@@ -1,0 +1,30 @@
+import os.path
+
+top = '.'
+out = 'build'
+
+def options(ctx):
+    ctx.load('pebble_sdk')
+
+def configure(ctx):
+    ctx.load('pebble_sdk')
+
+def build(ctx):
+    ctx.load('pebble_sdk')
+
+    binaries = []
+
+    for p in ctx.env.TARGET_PLATFORMS:
+        ctx.set_env(ctx.all_envs[p])
+        ctx.set_group(ctx.env.PLATFORM_NAME)
+
+        app_elf='{}/pebble-app.elf'.format(ctx.env.BUILD_DIR)
+        ctx.pbl_program(source=ctx.path.ant_glob('src/main/*.c'), target=app_elf)
+
+        worker_elf='{}/pebble-worker.elf'.format(ctx.env.BUILD_DIR)
+        ctx.pbl_worker(source=ctx.path.ant_glob('src/worker/*.c'), target=worker_elf)
+
+        binaries.append({'platform': p, 'app_elf': app_elf, 'worker_elf': worker_elf})
+
+    ctx.set_group('bundle')
+    ctx.pbl_bundle(binaries=binaries, js=ctx.path.ant_glob('src/js/**/*.js'))
